Rename rails engine and remove model definition when adding module
