+++ lib/devise/auth0/rails.rb
